modify oneToMany comments
