+++ java/spring/src/main/java/com/example/spring/relationships/oneToMany/Dojo.java
@@ -1,4 +1,4 @@
-package com.example.spring.relationships.onetomany;
+package com.example.spring.relationships.oneToMany;
 
 import com.example.spring.models.SuperModel;
 import jakarta.persistence.Entity;
@@ -12,9 +12,9 @@ import java.util.List;
 @Table(name = "dojos")
 public class Dojo extends SuperModel {
 
-    //* Here we change the one-to-one to the "@OneToMany" annotation, and don't include the cascade type.
+    //* "@OneToMany" doesn't include the cascade type, and precedes the collection.
     @OneToMany(mappedBy = "dojo", fetch = FetchType.LAZY)
 
-    //* And instead of only one field, we have a list of the other model.
+    //* Collection of linked objects of the other model.
     private List<Ninja> ninjas;
 }
